refactor: split up matrix initialization code a bit
