+++ src/Initializer/InitProcedure/InitModel.cpp
@@ -12,8 +12,10 @@
 #include "Config.h"
 #include "Equations/Datastructures.h"
 #include "Initializer/BasicTypedefs.h"
-#include "Initializer/CellLocalMatrices.h"
 #include "Initializer/MemoryManager.h"
+#include "Initializer/Model/BoundaryMappings.h"
+#include "Initializer/Model/CellLocalMatrices.h"
+#include "Initializer/Model/DynamicRuptureMatrices.h"
 #include "Initializer/ParameterDB.h"
 #include "Initializer/Parameters/ModelParameters.h"
 #include "Initializer/Parameters/SeisSolParameters.h"
