+++ src/Initializer/Model/DynamicRuptureMatrices.h
@@ -6,8 +6,8 @@
 // SPDX-FileContributor: Author lists in /AUTHORS and /CITATION.cff
 // SPDX-FileContributor: Carsten Uphoff
 
-#ifndef SEISSOL_SRC_INITIALIZER_CELLLOCALMATRICES_H_
-#define SEISSOL_SRC_INITIALIZER_CELLLOCALMATRICES_H_
+#ifndef SEISSOL_SRC_INITIALIZER_MODEL_DYNAMICRUPTUREMATRICES_H_
+#define SEISSOL_SRC_INITIALIZER_MODEL_DYNAMICRUPTUREMATRICES_H_
 
 #include "Geometry/MeshReader.h"
 #include "Initializer/Parameters/ModelParameters.h"
@@ -18,18 +18,6 @@
 #include "Memory/Tree/Backmap.h"
 
 namespace seissol::initializer {
-class EasiBoundary;
-/**
- * Computes the star matrices A*, B*, and C*, and solves the Riemann problems at the interfaces.
- **/
-void initializeCellLocalMatrices(const seissol::geometry::MeshReader& meshReader,
-                                 LTS::Storage& ltsStorage,
-                                 const ClusterLayout& clusterLayout,
-                                 const parameters::ModelParameters& modelParameters);
-
-void initializeBoundaryMappings(const seissol::geometry::MeshReader& meshReader,
-                                const EasiBoundary* easiBoundary,
-                                LTS::Storage& ltsStorage);
 
 void initializeDynamicRuptureMatrices(const seissol::geometry::MeshReader& meshReader,
                                       LTS::Storage& ltsStorage,
@@ -38,4 +26,4 @@ void initializeDynamicRuptureMatrices(const seissol::geometry::MeshReader& meshR
 
 } // namespace seissol::initializer
 
-#endif // SEISSOL_SRC_INITIALIZER_CELLLOCALMATRICES_H_
+#endif // SEISSOL_SRC_INITIALIZER_MODEL_DYNAMICRUPTUREMATRICES_H_
